backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pymongo>=4.10.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
openai>=1.55.0
httpx>=0.28.0
bcrypt>=4.2.0
email-validator>=2.2.0
